% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_npi.R
\name{as_npi}
\alias{as_npi}
\title{Convert a vector to an \code{npi} object}
\usage{
as_npi(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{These dots must be empty}
}
\value{
An S3 vector of class \verb{<npi>}
}
\description{
Converts a character or numeric vector of unvalidated NPIs into a validated \verb{<npi>} vector.
}
\examples{
as_npi(c("1234567891", 1234567891, 1234567891L, NA_character_))
}
