% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luhn.R
\name{check_digit}
\alias{check_digit}
\alias{last_digit}
\alias{examples}
\title{Luhn Check}
\usage{
check_digit(x)

last_digit(x)

examples()
}
\arguments{
\item{x}{a vector}
}
\value{
An integer vector
}
\description{
Helpers to construct the Luhn algorithm check
}
\examples{
x <- examples()
head(x, 20)

id <- last_digit(x)
head(id, 20)

x <- check_digit(x)
head(x, 20)

all(head(id, 20) == head(x, 20))
}
