% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luhn.R
\name{check_luhn}
\alias{check_luhn}
\alias{check_digit}
\alias{last_digit}
\alias{examples}
\title{Luhn Check}
\usage{
check_luhn(x, arg = rlang::caller_arg(x), call = rlang::caller_env())

check_digit(x)

last_digit(x)

examples()
}
\arguments{
\item{x}{a vector}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
An integer vector
}
\description{
Helpers to construct the Luhn algorithm check
}
\examples{
x <- examples()
check_luhn(x)
}
