% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate}
\alias{generate}
\title{Generate random NPIs}
\usage{
generate(x, prop_na = 0L)
}
\arguments{
\item{x}{integer, number of NPIs to generate}

\item{prop_na}{double; proportion of NAs to insert}
}
\value{
An integer vector of unvalidated NPIs
}
\description{
Creates an integer vector of unvalidated NPIs.
}
\examples{
generate(100, 0.1)
}
