% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npi.R
\name{npi}
\alias{npi}
\alias{is_npi}
\title{Create an \code{npi} vector}
\usage{
npi(x = integer())

is_npi(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{npi()}: An integer vector.
\item For \code{is_npi()}: An object to test.
}}
}
\value{
An S3 vector of class \code{npi}
}
\description{
This creates an integer vector of valid NPIs.
}
\examples{
x <- generate(10)
x
is_npi(x)

x <- npi(x)
x
is_npi(x)

data.frame(x)

tibble::tibble(x)

}
