% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npis-package.R
\docType{package}
\name{npis-package}
\alias{npis}
\alias{npis-package}
\title{npis: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/npis}
  \item \url{https://andrewallenbruce.github.io/npis/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/npis/issues}
}

}
\author{
\strong{Maintainer}: Andrew Allen Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
