% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_ptype_cast.R
\name{vctrs-methods}
\alias{vctrs-methods}
\alias{format.npi}
\alias{vec_ptype_abbr.npi}
\alias{vec_ptype_full.npi}
\alias{vec_ptype2.npi}
\alias{vec_ptype2.npi.npi}
\alias{vec_ptype2.integer.npi}
\alias{vec_ptype2.npi.integer}
\alias{vec_cast.npi}
\alias{vec_cast.npi.npi}
\alias{vec_cast.npi.integer}
\alias{vec_cast.integer.npi}
\title{Internal vctrs methods}
\usage{
\method{format}{npi}(x, ...)

\method{vec_ptype_abbr}{npi}(x, ...)

\method{vec_ptype_full}{npi}(x, ...)

\method{vec_ptype2}{npi}(x, y, ...)

\method{vec_ptype2}{npi.npi}(x, y, ...)

\method{vec_ptype2}{integer.npi}(x, y, ...)

\method{vec_ptype2}{npi.integer}(x, y, ...)

\method{vec_cast}{npi}(x, to, ...)

\method{vec_cast}{npi.npi}(x, to, ...)

\method{vec_cast}{npi.integer}(x, to, ...)

\method{vec_cast}{integer.npi}(x, to, ...)
}
\description{
These are internal vctrs methods.
}
\keyword{internal}
