% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npi.R
\name{as_npi}
\alias{as_npi}
\title{Convert to an \code{npi} vector}
\usage{
as_npi(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{empty dots}
}
\value{
An S3 vector of class \code{npi}
}
\description{
This creates an integer vector of valid NPIs.
}
\examples{
as_npi(c("1234567891", 1234567891, 1234567891L, NA_character_))
}
