% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate}
\alias{generate}
\title{Generate random NPIs}
\usage{
generate(x)
}
\arguments{
\item{x}{integer, number of NPIs to generate}
}
\value{
An integer vector of unvalidated NPIs
}
\description{
This creates an integer vector of unvalidated NPIs.
}
\examples{
generate(100)
}
