% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate}
\alias{generate}
\title{Generate random NPIs}
\usage{
generate(x = 100L, prop = 0.1, syntactic = TRUE)
}
\arguments{
\item{x}{integer, number of NPIs to generate}

\item{prop}{double; proportion of NAs to insert}

\item{syntactic}{boolean; technically valid NPIs}
}
\value{
An integer vector of unvalidated NPIs
}
\description{
Creates an integer vector of unvalidated NPIs.
}
\examples{
generate(100)
}
