% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R, R/npi.R
\name{generate_npi}
\alias{generate_npi}
\alias{npi}
\alias{new_npi}
\alias{is_npi}
\title{National Provider Identifier}
\usage{
generate_npi(x)

npi(x = integer())

new_npi(x = integer())

is_npi(x)
}
\arguments{
\item{x}{integer vector}
}
\value{
An S3 vctr of class \code{npi}
}
\description{
A National Provider Identifier (NPI) is...
}
\examples{
new_npi()
npi()
is_npi(npi())

x <- generate_npi(10)
npi(x)
data.frame(x)
tibble::tibble(x)
}
