% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R, R/npi.R
\name{generate_npi}
\alias{generate_npi}
\alias{npi}
\alias{new_npi}
\alias{is_npi}
\title{\code{npi} vector}
\usage{
generate_npi(x)

npi(x = integer())

new_npi(x = integer())

is_npi(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{npi()}: An integer vector.
\item For \code{is_npi()}: An object to test.
}}
}
\value{
An S3 vector of class \code{npi}
}
\description{
This creates an integer vector of valid NPIs.
}
\examples{
new_npi()
npi()
is_npi(npi())

x <- npi(generate_npi(10))
data.frame(x)
tibble::tibble(x)
}
