% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npis-package.R
\docType{package}
\name{npis-package}
\alias{npis}
\alias{npis-package}
\title{npis: Vectorized NPIs for R}
\source{
\href{https://www.cms.gov/Regulations-and-Guidance/Administrative-Simplification/NationalProvIdentStand/Downloads/NPIcheckdigit.pdf}{PDF: Requirements for National Provider Identifier (NPI) and NPI Check Digit}
}
\description{
An \href{https://github.com/r-lib/vctrs/}{S3 vctrs class} representing a
\strong{National Provider Identifier}, a unique 10-digit identification number
assigned to healthcare providers in the United States. This identifier is
crucial for billing and administrative transactions as mandated by the
Health Insurance Portability and Accountability Act (HIPAA).
\subsection{Requirements for the NPI}{
\enumerate{
\item The NPI must meet the specifications detailed in the \emph{Final Rule for the}
\emph{Standard Unique Health Identifier for Health Care Providers} (\href{https://www.federalregister.gov/documents/2004/01/23/04-1149/hipaa-administrative-simplification-standard-unique-health-identifier-for-health-care-providers}{69 FR 3434}),
published on January 23, 2004.
\item The NPI must consist of 9 numeric digits followed by one numeric check digit.
\item The NPI must not have embedded intelligence.
\item The NPI format and check digit calculation must be compatible with the
card issuer identifier on a standard health identification card. The
card standard was developed by NCITS, which is accredited by ANSI. Each
card issuer identifier used on a standard health identification card must
be unique. Therefore the NPI must be generated in a manner that will avoid
collisions with other identifiers that might be used as card issuer
identifiers on standard health identification cards. In order to avoid such
collisions, NPIs will initially be issued with a first digit of 1 or 2.
These digits will not be used as the first digits for other card issuer
identifiers. Use of other first digits must be coordinated with the use of first
digits by the standard health plan identifier, when adopted.
\item Within the constraints of No. 4. above, the NPI generation must use a
scattering algorithm that has the capability to use all possible numeric
combinations beginning with a 1 or 2.
\item Each NPI generated must be unique, without requiring database access for verification.
}
}

\subsection{Requirements for the NPI Check Digit}{

The NPI check digit is calculated using the \href{https://en.wikipedia.org/wiki/Luhn_algorithm}{Luhn algorithm}
for computing the modulus 10 "double-add-double" check digit, a recognized
ISO standard (\href{https://en.wikipedia.org/wiki/ISO/IEC_7812}{ISO/IEC 7812}).

When an NPI is used as a card issuer identifier on a standard health identification
card, it is prefixed by \code{80840}, in which \code{80} indicates \strong{Health Applications}
and \code{840} indicates \strong{United States}.

The check digit will always be calculated as if the prefix is present. This is
accomplished by simply adding the constant \code{24} in the check digit calculation
when the NPI is used without the prefix.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/npis}
  \item \url{https://andrewallenbruce.github.io/npis/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/npis/issues}
}

}
\author{
\strong{Maintainer}: Andrew Allen Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
