% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npis-package.R
\docType{package}
\name{npis-package}
\alias{npis}
\alias{npis-package}
\title{npis: Vectorized NPIs for R}
\source{
\href{https://www.cms.gov/Regulations-and-Guidance/Administrative-Simplification/NationalProvIdentStand/Downloads/NPIcheckdigit.pdf}{PDF: Requirements for National Provider Identifier (NPI) and NPI Check Digit}
}
\description{
An \href{https://github.com/r-lib/vctrs/}{S3 vctrs class} representing a
\strong{National Provider Identifier}, a unique 10-digit identification number
assigned to healthcare providers in the United States, as mandated by the
\href{https://en.wikipedia.org/wiki/Health_Insurance_Portability_and_Accountability_Act}{Health Insurance Portability and Accountability Act (HIPAA)}.
\subsection{NPI Requirements}{
\enumerate{
\item The NPI must meet the specifications detailed in the \strong{Final Rule for the}
\strong{Standard Unique Health Identifier for Health Care Providers} (\href{https://www.federalregister.gov/documents/2004/01/23/04-1149/hipaa-administrative-simplification-standard-unique-health-identifier-for-health-care-providers}{69 FR 3434}),
published 2004-01-23.
\item The NPI must consist of 9 numeric digits followed by 1 numeric check digit.
\item The NPI must not have embedded intelligence.
\item The NPI format and check digit calculation must be compatible with the
card issuer identifier on a standard health identification card. As each
card issuer identifier must be unique, the NPI must avoid collisions with
other possible card issuer identifiers. NPIs will therefore initially be
issued with a first digit of 1 or 2. These digits will be exclusive to
the NPI.
\item Within the constraints of (4), NPI generation must employ a scattering
algorithm that uses all possible numeric combinations beginning with a 1 or 2.
\item Each NPI generated must be unique, without requiring database access for verification.
}
}

\subsection{NPI Check Digit}{

The NPI check digit is calculated using the \href{https://en.wikipedia.org/wiki/Luhn_algorithm}{Luhn algorithm}
for computing the modulus 10 "double-add-double" check digit, a recognized
ISO standard (\href{https://en.wikipedia.org/wiki/ISO/IEC_7812}{ISO/IEC 7812}).

When an NPI is used as a card issuer identifier, it is prefixed by \code{80840},
in which \code{80} indicates \strong{Health Applications} and \code{840} indicates \strong{United States}.

The check digit will always be calculated as if the prefix is present. This is
accomplished by simply adding the constant \code{24} in the check digit calculation
when the NPI is used without the prefix.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/npis}
  \item \url{https://andrewallenbruce.github.io/npis/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/npis/issues}
}

}
\author{
\strong{Maintainer}: Andrew Allen Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
